+++ RELEASENOTES.md
@@ -6,6 +6,7 @@ AL-Go now offers a dataexplorer dashboard to get started with AL-Go telemetry. A
 
 - Issue 1770 Wrong type of _projects_ setting in settings schema
 - Issue 1787: Publish to Environment from PR fails in private repos
+- Issue 1512: Throw a warning before trying to publish symbols packages
 
 ### Add custom jobs to AL-Go workflows